Trait sp_executor::ExecutorApi

pub trait ExecutorApi<Block: BlockT>: Core<Block> {

    fn submit_execution_receipt_unsigned(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        opaque_execution_receipt: OpaqueExecutionReceipt
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_execution_receipt_unsigned_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        opaque_execution_receipt: OpaqueExecutionReceipt
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_transaction_bundle_unsigned(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        opaque_bundle: OpaqueBundle
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_transaction_bundle_unsigned_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        opaque_bundle: OpaqueBundle
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_fraud_proof_unsigned(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        fraud_proof: FraudProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_fraud_proof_unsigned_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        fraud_proof: FraudProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_bundle_equivocation_proof_unsigned(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        bundle_equivocation_proof: BundleEquivocationProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_bundle_equivocation_proof_unsigned_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        bundle_equivocation_proof: BundleEquivocationProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_invalid_transaction_proof_unsigned(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        invalid_transaction_proof: InvalidTransactionProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn submit_invalid_transaction_proof_unsigned_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        invalid_transaction_proof: InvalidTransactionProof
    ) -> Result<Option<()>, ApiError> { ... }

    fn extract_bundles(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        extrinsics: Vec<OpaqueExtrinsic>
    ) -> Result<Vec<OpaqueBundle>, ApiError> { ... }

    fn extract_bundles_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        extrinsics: Vec<OpaqueExtrinsic>
    ) -> Result<Vec<OpaqueBundle>, ApiError> { ... }

    fn extrinsics_shuffling_seed(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }

    fn extrinsics_shuffling_seed_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }

}

API necessary for executor pallet.

Provided methods

fn submit_execution_receipt_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
opaque_execution_receipt: OpaqueExecutionReceipt
) -> Result<Option<()>, ApiError>

Submits the execution receipt via an unsigned extrinsic.

fn submit_execution_receipt_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
opaque_execution_receipt: OpaqueExecutionReceipt
) -> Result<Option<()>, ApiError>

Submits the execution receipt via an unsigned extrinsic.

fn submit_transaction_bundle_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
opaque_bundle: OpaqueBundle
) -> Result<Option<()>, ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_transaction_bundle_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
opaque_bundle: OpaqueBundle
) -> Result<Option<()>, ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_fraud_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
fraud_proof: FraudProof
) -> Result<Option<()>, ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_fraud_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
fraud_proof: FraudProof
) -> Result<Option<()>, ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_bundle_equivocation_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
bundle_equivocation_proof: BundleEquivocationProof
) -> Result<Option<()>, ApiError>

Submits the bundle equivocation proof via an unsigned extrinsic.

fn submit_bundle_equivocation_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
bundle_equivocation_proof: BundleEquivocationProof
) -> Result<Option<()>, ApiError>

Submits the bundle equivocation proof via an unsigned extrinsic.

fn submit_invalid_transaction_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
invalid_transaction_proof: InvalidTransactionProof
) -> Result<Option<()>, ApiError>

Submits the invalid transaction proof via an unsigned extrinsic.

fn submit_invalid_transaction_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
invalid_transaction_proof: InvalidTransactionProof
) -> Result<Option<()>, ApiError>

Submits the invalid transaction proof via an unsigned extrinsic.

fn extract_bundles(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsics: Vec<OpaqueExtrinsic>
) -> Result<Vec<OpaqueBundle>, ApiError>

Extract the bundles from extrinsics in a block.

fn extract_bundles_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsics: Vec<OpaqueExtrinsic>
) -> Result<Vec<OpaqueBundle>, ApiError>

Extract the bundles from extrinsics in a block.

fn extrinsics_shuffling_seed(
&self,
__runtime_api_at_param__: &BlockId<Block>,
header: Block::Header
) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.

fn extrinsics_shuffling_seed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
header: Block::Header
) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.