Trait sp_executor::ExecutorApi

pub trait ExecutorApi<Block: BlockT, SecondaryHash: Encode + Decode>: Core<Block> {

    fn submit_execution_receipt_unsigned(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        execution_receipt: SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>
    ) -> Result<(), ApiError> { ... }
    fn submit_execution_receipt_unsigned_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        execution_receipt: SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>
    ) -> Result<(), ApiError> { ... }
    fn submit_transaction_bundle_unsigned(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        opaque_bundle: SignedOpaqueBundle
    ) -> Result<(), ApiError> { ... }
    fn submit_transaction_bundle_unsigned_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        opaque_bundle: SignedOpaqueBundle
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        fraud_proof: FraudProof
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        fraud_proof: FraudProof
    ) -> Result<(), ApiError> { ... }
    fn submit_bundle_equivocation_proof_unsigned(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        bundle_equivocation_proof: BundleEquivocationProof
    ) -> Result<(), ApiError> { ... }
    fn submit_bundle_equivocation_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        bundle_equivocation_proof: BundleEquivocationProof
    ) -> Result<(), ApiError> { ... }
    fn submit_invalid_transaction_proof_unsigned(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        invalid_transaction_proof: InvalidTransactionProof
    ) -> Result<(), ApiError> { ... }
    fn submit_invalid_transaction_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        invalid_transaction_proof: InvalidTransactionProof
    ) -> Result<(), ApiError> { ... }
    fn extract_bundles(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<OpaqueBundle>, ApiError> { ... }
    fn extract_bundles_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<OpaqueBundle>, ApiError> { ... }
    fn extract_receipts(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>>, ApiError> { ... }
    fn extract_receipts_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>>, ApiError> { ... }
    fn extract_fraud_proofs(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<FraudProof>, ApiError> { ... }
    fn extract_fraud_proofs_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<Vec<FraudProof>, ApiError> { ... }
    fn extrinsics_shuffling_seed(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }
    fn extrinsics_shuffling_seed_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }
    fn execution_wasm_bundle(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Cow<'static, [u8]>, ApiError> { ... }
    fn execution_wasm_bundle_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<Cow<'static, [u8]>, ApiError> { ... }
    fn executor_id(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<ExecutorId, ApiError> { ... }
    fn executor_id_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<ExecutorId, ApiError> { ... }
    fn best_execution_chain_number(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn best_execution_chain_number_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }

}

API necessary for executor pallet.

Provided Methods

fn submit_execution_receipt_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
execution_receipt: SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>
) -> Result<(), ApiError>

Submits the execution receipt via an unsigned extrinsic.

fn submit_execution_receipt_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
execution_receipt: SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>
) -> Result<(), ApiError>

Submits the execution receipt via an unsigned extrinsic.

fn submit_transaction_bundle_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
opaque_bundle: SignedOpaqueBundle
) -> Result<(), ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_transaction_bundle_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
opaque_bundle: SignedOpaqueBundle
) -> Result<(), ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_fraud_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
fraud_proof: FraudProof
) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_fraud_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
fraud_proof: FraudProof
) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_bundle_equivocation_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
bundle_equivocation_proof: BundleEquivocationProof
) -> Result<(), ApiError>

Submits the bundle equivocation proof via an unsigned extrinsic.

fn submit_bundle_equivocation_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
bundle_equivocation_proof: BundleEquivocationProof
) -> Result<(), ApiError>

Submits the bundle equivocation proof via an unsigned extrinsic.

fn submit_invalid_transaction_proof_unsigned(
&self,
__runtime_api_at_param__: &BlockId<Block>,
invalid_transaction_proof: InvalidTransactionProof
) -> Result<(), ApiError>

Submits the invalid transaction proof via an unsigned extrinsic.

fn submit_invalid_transaction_proof_unsigned_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
invalid_transaction_proof: InvalidTransactionProof
) -> Result<(), ApiError>

Submits the invalid transaction proof via an unsigned extrinsic.

fn extract_bundles(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<OpaqueBundle>, ApiError>

Extract the bundles from the given extrinsics.

fn extract_bundles_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<OpaqueBundle>, ApiError>

Extract the bundles from the given extrinsics.

fn extract_receipts(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_receipts_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<SignedExecutionReceipt<NumberFor<Block>, Block::Hash, SecondaryHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_fraud_proofs(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<FraudProof>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn extract_fraud_proofs_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsics: Vec<Block::Extrinsic>
) -> Result<Vec<FraudProof>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn extrinsics_shuffling_seed(
&self,
__runtime_api_at_param__: &BlockId<Block>,
header: Block::Header
) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.

fn extrinsics_shuffling_seed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
header: Block::Header
) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.

fn execution_wasm_bundle(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Cow<'static, [u8]>, ApiError>

WASM bundle for execution runtime.

fn execution_wasm_bundle_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Cow<'static, [u8]>, ApiError>

WASM bundle for execution runtime.

fn executor_id(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<ExecutorId, ApiError>

Returns the authority id of current executor.

fn executor_id_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<ExecutorId, ApiError>

Returns the authority id of current executor.

fn best_execution_chain_number(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn best_execution_chain_number_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn oldest_receipt_number(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn oldest_receipt_number_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn maximum_receipt_drift(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.

fn maximum_receipt_drift_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.